refactor: BrowseViewCollectionLayout 로직 분리
